port changes from dotnet/docs
